Move fetcher_optimization.py to fetching package
