+++ src/fetchers/djinni/fetcher.py
@@ -6,7 +6,7 @@ from playwright.async_api import async_playwright
 from src.utils.convert_bool import str_to_bool
 from src.fetchers.djinni.pagination import build_paginated_url
 from logs.logger import logger
-from src.utils.fetcher_optimization import block_resources
+from src.utils.fetching.fetcher_optimization import block_resources
 
 load_dotenv()
