+++ src/fetchers/pracuj/fetcher.py
@@ -9,7 +9,7 @@ from playwright.async_api import async_playwright
 
 from logs.logger import logger
 from src.utils.convert_bool import str_to_bool
-from src.utils.fetcher_optimization import block_pracuj_resources
+from src.utils.fetching.fetcher_optimization import block_pracuj_resources
 
 load_dotenv()
 PRACUJ_HEADLESS = str_to_bool(os.getenv("PRACUJ_HEADLESS", "false"))
