+++ src/fetchers/robota_ua/fetcher.py
@@ -7,7 +7,7 @@ from playwright.async_api import async_playwright
 from src.fetchers.robota_ua.pagination import click_next_page
 from logs.logger import logger
 from src.utils.convert_bool import str_to_bool
-from src.utils.fetcher_optimization import block_resources
+from src.utils.fetching.fetcher_optimization import block_resources
 
 load_dotenv()
